Exception handling - delete

+++ src/main/java/com/LucasRibasCardoso/course/services/UserService.java
@@ -2,10 +2,13 @@ package com.LucasRibasCardoso.course.services;
 
 import com.LucasRibasCardoso.course.entities.User;
 import com.LucasRibasCardoso.course.repositories.UserRepository;
+import com.LucasRibasCardoso.course.services.exceptions.DatabaseException;
 import com.LucasRibasCardoso.course.services.exceptions.ResourceNotFoundException;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 
@@ -29,7 +32,19 @@ public class UserService {
   }
 
   public void delete(Long id) {
-    userRepository.deleteById(id);
+    if (!userRepository.existsById(id)) {
+      throw new ResourceNotFoundException(id);
+    }
+
+    try {
+      userRepository.deleteById(id);
+    }
+    catch (EmptyResultDataAccessException e) {
+      throw new ResourceNotFoundException(id);
+    }
+    catch (DataIntegrityViolationException e) {
+      throw new DatabaseException(e.getMessage());
+    }
   }
 
   public User update(Long id, User user) {
